+++ examples/urls.js
@@ -1,0 +1,4 @@
+const autoParse = require('..')
+
+console.log('URL:', autoParse('https://example.com', { parseUrls: true }))
+console.log('Path:', autoParse('./foo/bar', { parseFilePaths: true }))